+++ src/components/transfer/List.js
@@ -9,6 +9,9 @@ import Pagination from "react-js-pagination";
 import upArrow from '../../images/up.svg';
 import downArrow from '../../images/down.svg';
 import { API_BASE_URL } from '../../constants';
+import { IMG_BASE_URL } from '../../constants';
+import { confirmAlert } from 'react-confirm-alert'; // Import
+import 'react-confirm-alert/src/react-confirm-alert.css'; // Import css
 
 class List extends Component {
 
@@ -59,16 +62,31 @@ class List extends Component {
     }
 
     async remove(id) {
-      await fetch(`${API_BASE_URL}/block/${id}`, {
-        method: 'DELETE',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json'
-        }
-      }).then(() => {
-        let updatedData = [...this.state.data].filter(i => i.id !== id);
-        this.setState({data: updatedData});
-      });
+      confirmAlert({
+        title: 'Confirm to submit',
+        message: 'Are you sure to delete this.',
+        buttons: [
+          {
+            label: 'Yes',
+            onClick: () => {
+              fetch(`${API_BASE_URL}/block/${id}`, {
+                method: 'DELETE',
+                headers: {
+                  'Accept': 'application/json',
+                  'Content-Type': 'application/json'
+                }
+              }).then(() => {
+                let updatedData = [...this.state.data].filter(i => i.id !== id);
+                this.setState({data: updatedData});
+              });
+            }
+          },
+          {
+            label: 'No',
+            onClick: () => alert('Click No')
+          }
+        ]
+      });      
     }
 
     render() {
@@ -82,11 +100,12 @@ class List extends Component {
         if (sortDirection === 'DESC'){
           sortImage = downArrow;
         }
-        const list = data.map(item => {            
+        const list = data.map(item => { 
+            let imgUrl = `${IMG_BASE_URL}/${item.image}`;           
             return <tr key={item.id}>
               <td style={{whiteSpace: 'nowrap'}}>{item.name}</td>        
               <td style={{whiteSpace: 'nowrap'}}>{item.source.name}</td>      
-              <td style={{whiteSpace: 'nowrap'}}><Image src={item.image} thumbnail /></td>
+              <td style={{whiteSpace: 'nowrap'}}><Image src={imgUrl} thumbnail /></td>
               <td>
                 <ButtonGroup>
                   <Button size="sm" variant="primary" href={"/transfer/" + item.id}>Edit</Button>
